// open-api-modules/dashboard-api-docs/model/apiResponseSearchAndPaginationResponseBank.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SearchAndPaginationResponseBank } from './searchAndPaginationResponseBank';


export interface ApiResponseSearchAndPaginationResponseBank { 
    result?: SearchAndPaginationResponseBank;
    errorCode?: string;
    message?: object;
    responseCode?: number;
}